Preparation for backend integration into plotter
Backends are now dynamically imported from the backend
directory, there is a backend option in the [plot] group of the
config which defines which backend name should be used for the figure.

For actual implementation, need to rewrite every option setting in the
plotter. As a transitional, can leave some options backend specific, but
should try to go backend independent on the plotter level.

Potentially, allow users to define plots only for a specific backend?
 - these would have names plot_"backendname" - could this be achieved
   with importlib?